Don't create build failed todo when build is retried
When a build is retried automatically, we close any open todos. However, we do
that _before_ creating a new build failed todo.

To solve this, we check if the build is retried before creating the todo. We
also ensure that the build _instance_ has the correct attribute set, without
needing to reload it from the database.