Remove Sherlock usage from the performance bar
Sherlock::Query generates a backtrace on every call to "new", which is
very slow. Formatter queries are also not displayed properly due to the
lack of "white-space: pre" in the CSS. We took a look at fixing this,
but the produced output is not really better than just displaying
queries on one line.

Fixes https://gitlab.com/gitlab-org/gitlab-ce/issues/39351