Fix broken Members link when relative URL root paths are used
Navigating to the "Members" section of a project would omit the
relative URL root path. Fix the redirection and eliminate the
need for a redirection in the first place.

Related Rails issue discussing omitting the leading slash in the
redirect paths: https://github.com/rails/rails/issues/21193

Fix other redirections failing to work with relative roots.