Spec the committer ChangeAccess check for merge requests
This tests the pre-receive hook that will be triggered when merging a
merge request by bringing the repository in a certain state.

For an FF merge, the `newrev` is going to be the last commit of the
merge request, so we can just pass the HEAD-sha of a branch that
applies cleanly.

For a normal merge, we need to create the merge commit, but not attach
it to any ref.