Fixup rspec tag misuse
In !14818 we added the ability to specify the example groups that should
run via the `-t` (tag) option to rspec.  It was assumed that the
inclusion filter generated `{:core => true, :mattermost => true}` would
run examples tagged :core OR :mattermost, when instead it means :core
AND :mattermost.

Here in order to assure at least some tests run, we specify that
the :mattermost tagged examples are the only ones we care about.