Fix test selection in Test::Integration::Mattermost
In ce3b81b953f we changed Test::Integration::Mattermost#perform to
consume a required positional parameter and pass on the rest of the
arguments to the superclass. There was an error in implementation here
and instead of splatting the files array to produce no additional
arguments, we passed in the default case an empty array.

This had the effect of calling rspec with no test files,
rather than expected default of 'qa/specs/features`

Here we correct that bug, and re-add the :core suite, as that was a poor
deuction in the investagtion for 74d6b8a211.