Fix access to the final page of todos
The todos page limit is 20, and both that and a user's pending todo count are
integers. Using integer division means that the result's floor will be taken,
defeating the point of the later call to `#ceil`.

So we need to convert one side of the division to a float first, otherwise the
last page won't be treated as available.