When a custom header logo is present, don't show GitLab type logo
In the new navigation header, when a user uploads a custom header
logo GitLab will still show the 'GitLab' type logo. This is often
undesirable as the user may want to upload a logo with their own
name present. This change will toggle the GitLab type logo if a
custom header logo is present.