PG HA: Rework headings to get unique links
As the flow of the PG HA documentation needs you to keep in mind several
roles, and visit them at multiple steps of the process, we end up with
mutiple headings called `Consul`, `PostgreSQL`, `Database`, and
`Pgbouncer`.

These headings become HTML anchors so a user can link to a particular
section of the document, but when needing to link to the Post-configure
of Consul you'd end up at the Consul section in user information.

Here we reword the headings.  They are less terse, but it allows for
deep-linking.

https://gitlab.com/gitlab-org/build/team-tasks/issues/30