Correct the host used for consul->pgbouncer pgpass
On failover the consul agent notifies the pgbouncer by running
`gitlab-ctl pgb-notify --newhost ADDRESS --user pgbouncer --hostuser gitlab-consul`.

For the documented configuration, this turns into a connection to the
pgbouncer service on 127.0.0.1, which may not be what was selected for
PGBOUNCER_HOST in the pgpass file, leading to authentication failures.