Merge alpha_database into database
As these instructions are now GA for 10.2, we move the contents of
alpha_database.md into database.md, as a new section `Configure using
Omnibus for High Availability`.  The old section heading is kept to
redirect the user to either the new guide, or the old version of the docs.

We move the architecture discussion up, to supplement the overview of
the install.