Improve output for extra queries in specs
Previously, this used `Array#-`, which would remove all queries that matches the
query text in the original set.

However, sometimes we have a problem with parameterised queries, where the query
text is identical both times, so we'd run a query N times instead of once, and
it would be hidden from the output.

Replace the logic to only remove a given query N times from the actual log,
where N is the number of times it appears in the expected log.