Add example configurations for PG HA
* Adds recommended example
* Adds minimal example

The recommended example is a 7-node configuration.  3x consul, 3x
database, 1x application.  Pgbouncer runs on the application node.

The minimal example is 4 nodes.  3x database, 1x application.  Pgbouncer
runs on the application node, we don't have a dedicated consul tier.

The style of these examples are copied from the existing redis examples.

Note, it is intended to supplement the preceding sections, it should not
be your only reference in setting up PG HA.