Tell the user they can skip ahead if they need to
At the end of the `Ensure GitLab is running` step, the user should have
a working setup and may be done with the guided tour.  It may be the
user has a problem, so instead needs to jump ahead to troubleshooting,
but the current structure is:

* Required information
* Initial node configuration
  * Consul
  * Database
  * Pgbouncer
  * Application
* Node post configuration
  * Consul
  * Database
  * Pgbouncer
  * Application
* Failover
* Restore
* Alternative configuration
* Troubleshooting

Here we just add a link to the end of 'Node post configuration / Application', rather than restructure.