Check if shard configuration is same across Geo nodes
This merge request will allow the primary to compare its storage configuration
against the secondaries. Not that the status only works if you view the admin
page on the primary, since the secondaries don't have access to the primary.

How it works:

1. The primary will populate the GeoNode index page with a JSON of the current configuration
2. The primary will download the latest configuration with the `/api/v4/geo/status` endpoint
3. The JavaScript on the frontend will do a simple equality comparison between the arrays

Since the shard configuration is currently stored in `gitlab.yml`, I've
purposely tried avoided storing the shard configuration in the database to
avoid multiple sources of the truth.

Closes #3243