Allow sidekiq to react to becoming a Geo primary or secondary without a restart
Sidekiq runs a set of cron jobs. This set differs according to whether the node
is a Geo primary or secondary, or if Geo is disabled. Previously, the state was
determined once, when sidekiq was started. Now, we add a cron job that runs the
same logic each minute.

To reduce churn, the enable/disable logic now only changes jobs if they are in
the wrong state.