Warm up the app so it's fast for the 1st request
This would make the application considered ready much slower,
but when it's ready, then it's really ready. Before this change,
it claims to be ready, but it's annoyingly slow for the first
request with GDK. It's 100% 502 for me, for the first request.

This shouldn't really affect production or so, because if it's
really ready, it should be blazingly fast, and it should not
slow things down too much.

The culprit here is probably `ActionController::Base.helpers.asset_path`
but this could make sure that anything else would load first, too.