Prevent cluster migration spec from interfering with later specs
If a later spec tries to set a field that is added to any of these models by a
later migration than the GCP clusters migration, it will fail because the
model's columns are cached.

We should always refer to models in migration specs by either their class in the
migration, or the table directly, using the `table` helper.