Fix stored XSS in code blocks
There were three things here:

1. Display math was broken.
2. <script> tags could be injected into code blocks with the language as `math`,
   `mermaid`, or `plantuml`.
3. <script> tags could be injected if Rouge threw an exception, for whatever
   reason.

This fixes all of those by always using the same code path for 'standard'
highlighting and 'special' languages (mathematics, Mermaid, and PlantUML), and
skipping the filter entirely if Rouge fails on a retry with the plain text
filter. It also adds specs for KaTeX and Mermaid rendering.