adapt the models that uses mounted uploaders
`ObjectStorage::Concern` adds delegation to the model that
mount an uploader that uses `RecordsUploads::Concern` to track
its uploads. The basic implementation is to use the uploader
to build paths that can be used to fetch back the related `Upload`
instance so it can be used to build the uploader back (which is
needed to build the correct file URL).