Fix LDAP attributes sync ssh keys boolean value bug
We tell users to set `sync_ssh_keys` to `false` if they don't
intend to use the feature. Strangely, `true` is not how you
enable it, rather you give it an LDAP attribute name. The
default value for `sync_ssh_keys` is `nil` which doesn't break
`Gitlab::LDAP::Person.ldap_attributes`, but `false` does. This
rejects `false` *and* `true` for good measure and adds an EE-only
test to ensure it doesn't regress.

Closes #4662