Allow to find epics by the given time frame
On the epics roadmap page (see #3559) we need to display epics
in the given time frame. This patch adds an additional filter to
the Epic finder which allows to filter only epics which overlap
with the given time frame, epics which do not have start and end date
set are intentionally ignored.

This filter is not used anywhere yet, but it will be used in the MR for #3559.
This patch is a separate MR to avoid overgrowing and mixing too many
things together in the frontend patch for #3559.

Related #3559